正在解决数据接受问题，已有成功案例--沙俊霖

+++ modules/master_machine/seasky_protocol.h
@@ -110,7 +110,7 @@ typedef __packed struct
 	float linearx;
 	float linery;
 	float angularz;
-	uint8_t blank;
+	uint16_t blank;
 	uint8_t check_byte;
 	uint8_t frame_tail;
 }BUBING_CTRL; //32
